requirements.txt
Flask>=2.2.0
Werkzeug>=2.2.0
google-api-python-client>=2.60.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.0
Flask-Cors>=3.0.0
Flask-SQLAlchemy>=2.5.0
